No longer use os.Exit to quit application
It turns out that `os.Exit` quits the program without honoring any
`defer` calls.  It is now restricted to being called at the end of the
main function if an error has occured, or to exit the application after
printing the help text to the terminal.  All other instances of exiting
occurs by boiling up an error to the main function.